update blt and svelte page integration (bug bluetooth to fix)
BUG TO FIX: bug on requestDevice
